refactor(types): split wire schemas into per-zero-protocol modules
Mirror the upstream zero-protocol file layout one-to-one (direct
translation) instead of the monolithic push.ts/queries.ts:

- rename src/types/queries.ts -> custom-queries.ts
- add ast.ts, mutation-id.ts, mutation.ts, mutate-server.ts, query-server.ts
- push.ts now holds only the push-request/legacy types; the mutation and
  1.5+ response shapes (MutateResponse/QueryResponse) live in their own
  modules
- update server.ts, server-transaction.ts, README, and tests to match
