+++ src/types/ast.ts
@@ -1,0 +1,4 @@
+import * as Schema from "effect/Schema";
+
+/** TODO: Defined as unknown for now */
+export const Ast = Schema.Unknown;
