+++ tests/index.test.ts
@@ -33,8 +33,9 @@ import * as ZfxQuery from "effect-zero/query";
 import * as ZfxResult from "effect-zero/result";
 import * as ZfxServer from "effect-zero/server";
 import * as ZfxServerTransaction from "effect-zero/server-transaction";
-import { PushBody, PushParams, type PushResponse } from "effect-zero/types/push";
-import { TransformRequestMessage } from "effect-zero/types/queries";
+import { TransformRequestMessage } from "effect-zero/types/custom-queries";
+import { MutateParams, type MutateResponse } from "effect-zero/types/mutate-server";
+import { PushBody } from "effect-zero/types/push";
 import { nanoid } from "nanoid";
 import postgres from "postgres";
 import { WebSocket } from "undici";
@@ -295,22 +296,22 @@ const initZero = Effect.gen(function* () {
   return z;
 });
 
-let responses = Chunk.empty<PushResponse>();
+let responses = Chunk.empty<MutateResponse>();
 
 beforeAll(async () => {
   const pushRoute = HttpRouter.add(
     "POST",
     "/push",
     Effect.gen(function* () {
-      const params = yield* HttpRouter.schemaParams(PushParams);
+      const params = yield* HttpRouter.schemaParams(MutateParams);
       const payload = yield* HttpServerRequest.schemaBodyJson(PushBody);
       // `handleMutate` returns the upstream push response already in wire format, so no encoding step.
       // No auth in this test harness, so the client is logged out (`userId: undefined`).
       const result = yield* ZfxServer.handleMutate({
         transaction: serverTransaction,
         mutators: serverMutators,
-        params,
-        request: payload,
+        query: params,
+        body: payload,
         userId: undefined,
       });
 
@@ -335,7 +336,7 @@ beforeAll(async () => {
     "/query",
     Effect.gen(function* () {
       const payload = yield* HttpServerRequest.schemaBodyJson(TransformRequestMessage);
-      const response = yield* ZfxServer.handleQuery({ queries, schema, payload, userId: undefined });
+      const response = yield* ZfxServer.handleQuery({ queries, schema, body: payload, userId: undefined });
       return yield* HttpServerResponse.json(response);
     }).pipe(
       Effect.catchCause(
@@ -366,7 +367,7 @@ beforeAll(async () => {
 });
 
 beforeEach(() => {
-  responses = Chunk.empty<PushResponse>();
+  responses = Chunk.empty<MutateResponse>();
 });
 
 test("server is running", async () => {
@@ -378,8 +379,8 @@ test("handleMutate has no extra requirements", () => {
   const effect = ZfxServer.handleMutate({
     transaction: serverTransaction,
     mutators: serverMutators,
-    params: {} as any,
-    request: {} as any,
+    query: {} as any,
+    body: {} as any,
     userId: undefined,
   });
 
@@ -432,8 +433,8 @@ test("mutator requirements should propagate", () => {
   const serverEffect = ZfxServer.handleMutate({
     transaction: serverTransaction,
     mutators,
-    params: {} as any,
-    request: {} as any,
+    query: {} as any,
+    body: {} as any,
     userId: undefined,
   });
 
@@ -760,7 +761,7 @@ it.live(
   "unsupported push version returns a top-level PushFailed",
   Effect.fn(function* () {
     // The push-version check runs before any DB/client logic, so dummy params + an empty batch suffice.
-    const params: PushParams = { schema: "public", appID: "zero" };
+    const params: MutateParams = { schema: "public", appID: "zero" };
     const body: PushBody = {
       clientGroupID: nanoid(),
       mutations: [],
@@ -772,8 +773,8 @@ it.live(
     const result = yield* ZfxServer.handleMutate({
       transaction: serverTransaction,
       mutators: serverMutators,
-      params,
-      request: body,
+      query: params,
+      body,
       userId: undefined,
     });
     expect(result).toMatchObject({ kind: "PushFailed", origin: "server", reason: "unsupportedPushVersion" });
